requirements.txt
p5.aiogoldsrcrcon
python-telegram-bot
vk_api
asyncio~=3.4.3
mysql-connector-python~=8.0.31